+++ particle-actors/src/vm_pool.rs
@@ -117,7 +117,7 @@ fn create_vm(
             aquamarine_wasm_path: config.air_interpreter,
             particle_data_store: config.particles_dir,
             call_service: host_closure(),
-            logging_mask: i64::max_value(),
+            logging_mask: i32::max_value(),
         };
         let vm = AquamarineVM::new(config)?;
         if let Some(waker) = waker.read().as_ref() {
